requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cleanup==9.0.0
django-environ==0.12.0
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
psycopg==3.3.2
sqlparse==0.5.5
typing_extensions==4.15.0
